Update requirements for dev environment
Following the install steps caused an initial issue that required a
version bump for `jupyter_server` up to 2.0.0. This had a knock-on
effect of requiring further bumps in one other packages: tornado.

This commit relaxes the rules in the requirements.txt and setup.py files
wherever `jupyter_server` or `tornado` are mentioned. It furthermore
removes three lines from the `grader_service/main.py` that was calling a
routine to remove handlers with the express purpose of preventing
duplicate logging. As tornado was being used for logging, it seems like
there may have been an api change between our version bumps that may
have cased this function to fail.

This commit closes issue #65

Signed-off-by: Adam McCartney <adam@mur.at>