Part 2 of the git lookup refactoring saga
Note: this commit was added again after rebasing remotes.

Rolled back changes introduced by sha bea4147b. We had not been running
the test suite while introducing the changes, so we needed to roll back
in order to add them (again) incrementally. This commit does not
introduce anything substantial beyond what was already started in the
previous commit. It just guarantees that the test suite will pass
(at least locally, let's see what the pipeline has to say!)

Most of the changes focus on the 'handlers/git/server' module,
specifically on the GitBaseHandler class and, even more specifically, on
the 'GitBaseHandler.gitlookup()' method.

The changes mostly focus on introducing some dataclass objects to cut
down on the use of primitives and also a number of helper methods
that aim to make it a bit easier for us to read and reason about what
gitlookup is doing.

In addition to the changes mentioned above, a number of modifications
have been made to the 'tests/handlers/test_git' suite. These changes
involve removal of duplicate code with the introduction of a number of fixtures.
Also a number of additional return values were added to the mocks in
order to simulate what the real class would be doing.

Signed-off-by: Adam McCartney <adam@mur.at>