Part 3 of the git lookup refactoring saga
We've thrown out the use of StrEnum as it is python3.11 dependent and
probably a lazy man's enum solution. Instead we are using standard
enums, but constructing these in a more traditional way. The enum type
is used to construct a RepoTypeToken that we then use to parse the
request string.

Git lookup is still doing quite a lot. I __think__ it's safe enough to
say that it's main responsibility is to parse an incoming request in
order to determine what sort of git command needs to be run. In some
cases this causes the initialization of a new git repo.

For me currently the most opaque part of the function now from line 174
onwards, where it begins to call inherited routines in order to create
the git directory for a release.

Signed-off-by: Adam McCartney <adam@mur.at>