Part 4 of git lookup reafactoring saga
Note: untested code ahead!

Slight refactor of classes related to GitRepoBasePath. This commit
updates the prviously introduced classes used for representing our git
repositories within the context of GitBaseHandler. These edits use
slightly more python-like calls to the various `super.__init__()`
methods used in constructing the objects. Also we are introducing
respective `__repr__` methods. It seems like we are already representing
the git repo path objects as strings in various places -- specifically
where we are using `os.path.join(...)` to form the paths on the system.
We might as well do this explicitly rather than manually constructing
the representations within the context of every call to a method from
the `os.path` library.

Ideally I would like to clarify if these three classes are now able to
faithfully represent the types of directory trees that we want to create
for our git repositories.

Signed-off-by: Adam McCartney <adam@mur.at>