Part 5 of git lookup saga
Fix a couple of errors spotted by Josef after being introduced by me on
friday. Both happen in the context of the git lookup subroutine. The
first was that we were performing a check for the number of pathlets
that get created upon parsing a request. The bug was introduced because
we were checking for the total length of the list immediately after it
had been created from the request. The fix is that we should first
abbreviate the list. We also need to perform a small calculation to
determine the actual position of the last ocurrance of git in the list.

The second bug was a logic error:
we were using "if not ((True) or (True))" where we should have been
using "if not ((True) and (True))".

Signed-off-by: Adam McCartney <adam@mur.at>