Grader-service-migrate load config from file
Gives the option to load the config from the same file we use for
`grader-service`. It just parses the line containing the database url by
way of matching a regular expression. Closes issue #90. Adds a few
tests. It doesn't really have any proper checks in place for when config
is passed an invalid db url.