Turn back on strict compiler checking
Reintroduction of the `strict: true` compiler flag initially causes the
build of the lab extension to fail. It was complaining about possibly
passing an `undefined` type to the `enqueSnackbar`. This solution just
explicitly checks if the caught exception is an Error, then enques it,
otherwise is logs a message (unknown error) to the debug console.