Add tables to student / assignment view
This commit adds a table components as per the coursemanage
view, the lectures table is pretty much exactly the same as the
coursemanage/lecture table. The assignments table has some aspects that
make it unique to the student viewpoint. This commit gets all the icons
and that in place, does not yet implement navigation logic correctly.