Add functionality to buttons on assignment table
This commit adds functionality to the "reset" button on the assignment
table (the main student view of a lecture's assignments). It also
removes the "submit" button from this view. The reason for removing the
submit button is that it is arguably better placed within the detail
view of the assignment. The assumption is that a student rarely will
want to make a submissions without having some overview of the files
that are being submitted. By omitting the submit button, we reduce the
amount of state management needed for the assignment table view, reduce
load times for the view and reduce the amount of code duplication in our
code base.