Quickfix for deadline-wrapper + git error handling
Came across a bug in testing new paths feature: deadline was not getting
displayed when loading file view. Here is the fix.

Also included in this commit are a few addditions to the version
control handler that avoid type errors when handling exceptions.
We either invoking the __str__ method or calling the requisite attribute
for the error type.