Made tabpanels responsive
Cherry-picking this for the
feature/react-router-for-dataloading-components branch. The
feature/labextension/side-navigation branch has code in it from last
week that contains bugs that were since fixed in the merge base.

Restyled Overview tab, changed progress stepper to vertical one, added deadline component and most important
features of assignment as cards.