Fix: feedback generation failing with student notebooks (#288)
* fix: Improve logging when feedback generation fails

* chore: Remove an unused dataclass, fix some typos and type annotations

* fix: Fix a potential security issue when running Git commands in a subprocess

When running multiple commands with `sh -c ...`, it would be easier to construct
the variable "message" so that it contained some executable code.

* fix: Fix a typo in the log message after a successful autograde job

* fix: Do not try to generate feedback for notebooks created by students

If the instructor specifies '*.ipynb' in the whitelist of an assignment,
additional notebooks created by students are also copied over. The feedback
generator should not try to generate feedback for them - they are not in
the gradebook and an attempt to do so would fail.

Refs: #272

* fix: Improve error handling and logging in GitBaseHandler.git_reponse()