refactor: Pass git commands to subprocess.run as lists, not strings
We execute the git commands with variables potentially provided by users
(filenames, usernames, etc.). Passing them as strings and then splitting,
potentially could enable constructing these variables in such a way
so as to execute malicious commands.

An alternative would be to use `shlex.quote` with any variables,
but it seems more cumbersome.